+++ lua/jokers/ferretlady.lua
@@ -2,7 +2,9 @@ SMODS.Joker {
     atlas = "joker",
     key = "ferret",
     pos = { x = 6, y = 0 }, soul_pos = { x = 0, y = 0 },
-    set_badges = function(self, card, badges) badges[#badges+1] = tss_badges.Goob() end,
+    set_badges = function(self, card, badges) if (self.discovered) then
+		badges[#badges+1] = slimeutils.table_create_badge(tss_badges.Goob)
+	end end,
     discovered = true,
     rarity = 1,
     cost = 4,