+++ lua/jokers/flamenoble.lua
@@ -2,7 +2,10 @@ SMODS.Joker {
     atlas = "joker",
 	key = "flame",
 	pos = { x = 1, y = 0 }, soul_pos = { x = 0, y = 0 },
-    set_badges = function(self, card, badges) badges[#badges+1] = tss_badges.Quake() end,
+    set_badges = function(self, card, badges) if (self.discovered) then
+		badges[#badges+1] = slimeutils.table_create_badge(tss_badges.Quake)
+		badges[#badges+1] = slimeutils.table_create_badge(tss_badges.SO)
+	end end,
     discovered = true,
     rarity = 2,
     cost = 6,