+++ lua/jokers/goblin.lua
@@ -1,7 +1,7 @@
-SMODS.Joker{
+local bankroll = SMODS.Joker{
     atlas = 'joker',
-    key = 'giggles',
-    pos = { x = 8, y = 0 }, soul_pos = { x = 0, y = 0 },
+    key = 'goblin',
+    pos = { x = 1, y = 1 }, soul_pos = { x = 0, y = 0 },
     set_badges = function(self, card, badges) if (self.discovered) then
 		badges[#badges+1] = slimeutils.table_create_badge(tss_badges.Goob)
 	end end,
@@ -9,5 +9,5 @@ SMODS.Joker{
     unlocked = false,
     rarity = 2,
     cost = 6,
-    config = { extra = {} }
+    config = { extra = { } }
 }