+++ lua/jokers/roulette.lua
@@ -15,7 +15,10 @@ SMODS.Joker{
     atlas = 'joker',
     key = 'roulette',
     pos = { x = 5, y = 0 },
-    set_badges = function(self, card, badges) badges[#badges+1] = tss_badges.Xtra() end,
+    set_badges = function(self, card, badges) if (self.discovered) then
+		badges[#badges+1] = slimeutils.table_create_badge(tss_badges.Xtra)
+		badges[#badges+1] = slimeutils.table_create_badge(tss_badges.SO)
+	end end,
     discovered = true,
     rarity = 3,
     cost = 8,